feat: question3 progress
Question 3 is almost compleated however it has a bug in the implementation. It is not taking all the possible edges into consideration, only the ones starting at the current node. This could be changed by going through all the edges visible from the nodes in the stack which include the current node